Cleaning up comments around env vars
Also, move `getTemporaryDirectory` back to `actions-util`.
