+++ lib/actions-util.js
@@ -43,6 +43,13 @@ function getOptionalInput(name) {
     return value.length > 0 ? value : undefined;
 }
 exports.getOptionalInput = getOptionalInput;
+function getTemporaryDirectory() {
+    const value = process.env["CODEQL_ACTION_TEMP"];
+    return value !== undefined && value !== ""
+        ? value
+        : util_1.getRequiredEnvParam("RUNNER_TEMP");
+}
+exports.getTemporaryDirectory = getTemporaryDirectory;
 function getToolCacheDirectory() {
     const value = process.env["CODEQL_ACTION_TOOL_CACHE"];
     return value !== undefined && value !== ""
