+++ lib/analyze-action.js
@@ -40,7 +40,7 @@ async function run() {
             return;
         }
         const logger = logging_1.getActionsLogger();
-        config = await config_utils_1.getConfig(util.getTemporaryDirectory(), logger);
+        config = await config_utils_1.getConfig(actionsUtil.getTemporaryDirectory(), logger);
         if (config === undefined) {
             throw new Error("Config file could not be found at expected location. Has the 'init' action been called?");
         }
