+++ lib/autobuild-action.js
@@ -37,7 +37,7 @@ async function run() {
         if (!(await actions_util_1.sendStatusReport(await actions_util_1.createStatusReportBase("autobuild", "starting", startedAt)))) {
             return;
         }
-        const config = await config_utils.getConfig(util_1.getTemporaryDirectory(), logger);
+        const config = await config_utils.getConfig(actions_util_1.getTemporaryDirectory(), logger);
         if (config === undefined) {
             throw new Error("Config file could not be found at expected location. Has the 'init' action been called?");
         }
