+++ src/analyze-action.ts
@@ -67,7 +67,7 @@ async function run() {
       return;
     }
     const logger = getActionsLogger();
-    config = await getConfig(util.getTemporaryDirectory(), logger);
+    config = await getConfig(actionsUtil.getTemporaryDirectory(), logger);
     if (config === undefined) {
       throw new Error(
         "Config file could not be found at expected location. Has the 'init' action been called?"
